pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "elimu-research-assistant"
dynamic = ["version"]
description = "An intelligent research assistant for Kenyan educators to create localized, contextual educational content that bridges the context deficit in education."
readme = {file = "README.md", content-type = "text/markdown"}
license = {file = "LICENSE"}
authors = [
  {name = "Ashioya Jotham", email = "victorashioya960@gmail.com"}
]
keywords = [
  "education",
  "kenya",
  "ai",
  "research",
  "teaching",
  "localized-content",
  "curriculum",
  "pedagogy",
  "african-education",
  "context-aware",
  "educational-technology"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Development Status :: 4 - Beta",
  "Intended Audience :: Education",
  "Intended Audience :: End Users/Desktop",
  "Topic :: Education",
  "Topic :: Education :: Computer Aided Instruction (CAI)",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Text Processing :: Linguistic",
  "Topic :: Internet :: WWW/HTTP :: Indexing/Search"
]
requires-python = ">=3.9"
dependencies = [
  "click>=8.0.0",
  "requests>=2.25.0",
  "beautifulsoup4>=4.9.0",
  "html2text>=2020.1.16",
  "google-generativeai>=0.8.3",
  "python-dotenv>=0.19.0",
  "prompt_toolkit>=3.0.0",
  "rich>=10.0.0",
  "keyring>=23.0.0"
]

[project.urls]
"Source Code" = "https://github.com/ashioyajotham/elimu_research_assistant"
"Bug Tracker" = "https://github.com/ashioyajotham/elimu_research_assistant/issues"
"Documentation" = "https://github.com/ashioyajotham/elimu_research_assistant#readme"
"Educational Impact" = "https://github.com/ashioyajotham/elimu_research_assistant/wiki/Educational-Impact"

[project.scripts]
elimu = "cli:main"

[tool.setuptools.dynamic]
version = {file = "VERSION"}

[tool.setuptools.packages.find]
include = ["*"]

